--- app/src/main/java/com/kionavani/todotask/ui/Navigation.kt
@@ -1,9 +1,0 @@
-package com.kionavani.todotask.ui
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-object MainScreenNav
-
-@Serializable
-data class AddTaskScreenNav(val itemID: String?)
