+++ app/src/main/java/com/kionavani/todotask/ui/composable/MainScreen.kt
@@ -1,4 +1,4 @@
-package com.kionavani.todotask.ui.composable
+package com.kionavani.todotask.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -38,17 +38,15 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.kionavani.todotask.R
 import com.kionavani.todotask.data.Importance
 import com.kionavani.todotask.data.ToDoItem
-import com.kionavani.todotask.domain.LocalNavController
-import com.kionavani.todotask.ui.viewmodels.TodoViewModel
+import com.kionavani.todotask.ui.viewmodels.ToDoViewModel
 
 @Composable
-fun MainScreen(viewModel: TodoViewModel) {
+fun MainScreen(viewModel: ToDoViewModel) {
     val navController = LocalNavController.current
     val tasks by viewModel.todoItems.collectAsState()
 
@@ -57,9 +55,6 @@ fun MainScreen(viewModel: TodoViewModel) {
     }
     val getDeadlineDate = { item: ToDoItem ->
         item.deadlineDate?.let { viewModel.dateToString(it) }
-    }
-    val getDescription = { item: ToDoItem ->
-        viewModel.getDescWithEmoji(item)
     }
 
     var visibilityButtonState by remember {
@@ -141,8 +136,7 @@ fun MainScreen(viewModel: TodoViewModel) {
                     tasks
                 },
                 changeTaskState,
-                getDeadlineDate,
-                getDescription
+                getDeadlineDate
             )
         }
     }
@@ -153,8 +147,7 @@ fun MainScreen(viewModel: TodoViewModel) {
 fun TaskList(
     tasks: List<ToDoItem>,
     changeTaskState: (String, Boolean) -> Unit,
-    getDeadlineDate: (ToDoItem) -> String?,
-    getDescription: (ToDoItem) -> String
+    getDeadlineDate: (ToDoItem) -> String?
 ) {
     val navController = LocalNavController.current
 
@@ -169,7 +162,7 @@ fun TaskList(
             ),
     ) {
         items(tasks) { task ->
-            Task(item = task, changeTaskState, getDeadlineDate, getDescription)
+            Task(item = task, changeTaskState, getDeadlineDate)
 
         }
         item {
@@ -193,13 +186,10 @@ fun TaskList(
 fun Task(
     item: ToDoItem,
     changeTaskState: (String, Boolean) -> Unit,
-    getDeadlineDate: (ToDoItem) -> String?,
-    getDescription: (ToDoItem) -> String
+    getDeadlineDate: (ToDoItem) -> String?
 ) {
     val navController = LocalNavController.current
     val deadlineDate = getDeadlineDate(item)
-    val description = getDescription(item)
-
 
     Row(
         verticalAlignment = Alignment.CenterVertically,
@@ -215,7 +205,13 @@ fun Task(
                 .padding(start = 12.dp, top = 12.dp)
                 .weight(1f)
         ) {
-            ItemDescription(description, item.isCompleted)
+            Text(
+                text = item.taskDescription,
+                maxLines = 3,
+                overflow = TextOverflow.Ellipsis,
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onPrimary,
+            )
 
             if (deadlineDate != null) {
                 Text(
@@ -239,28 +235,6 @@ fun Task(
                 tint = MaterialTheme.colorScheme.onTertiary
             )
         }
-    }
-}
-
-@Composable
-fun ItemDescription(description: String, isCompleted: Boolean) {
-    if (isCompleted) {
-        Text(
-            text = description,
-            maxLines = 3,
-            overflow = TextOverflow.Ellipsis,
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onTertiary,
-            textDecoration = TextDecoration.LineThrough
-        )
-    } else {
-        Text(
-            text = description,
-            maxLines = 3,
-            overflow = TextOverflow.Ellipsis,
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onPrimary
-        )
     }
 }
